+++ packages/components/search-bar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.34](https://github.com/Synerise/synerise-design/compare/@synerise/ds-search-bar@0.2.33...@synerise/ds-search-bar@0.2.34) (2020-04-29)
+
+**Note:** Version bump only for package @synerise/ds-search-bar
+
+
+
+
+
 ## [0.2.33](https://github.com/Synerise/synerise-design/compare/@synerise/ds-search-bar@0.2.32...@synerise/ds-search-bar@0.2.33) (2020-04-27)
 
 **Note:** Version bump only for package @synerise/ds-search-bar